backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
requests==2.31.0
qrcode==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0
cryptography==41.0.7